+++ src/lib.rs
@@ -1,2 +1,3 @@
 pub mod units;
 pub mod utils;
+pub mod commands;
